requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Pillow==8.3.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
